biodyn class

A class that implement a biomass dynamic stock assessment model.

Details

The Class is intended to be used as part of an MSE and includes methods for diagnostics, calculating reference points and other quantities used when providing management advice.

Slots

name

A character with the name of the stock

desc

A character providing a fuller description of the object

range

A numeric vector containing the quant and year ranges

model

A factor giving name of production function, for now this is only `pellat`

obj

factor that determines the objective function type -LL or LAV

catch

An FLQuant with total catch by year

index

An FLQuants with indices year

stock

An FLQuant which will hold the estimated stock by year

control

An FLPar which sets initial guess (val) and bounds (min and max) for each parameter. The phase allows a parameter to be fixed if less <0 and for paramters to be estimated sequentially

hcr

A data.frame with harvest control rule options

priors

An array which sets penalties for parameters

params

An FLPar with parameter estmates

vcov

An FLPar with the covariance matrix of the parameter estimates

hessian

An FLPar with the hessian of the estimated parameters

ref

A numeric with parameters for estimating mng quantities

mng

FLPar with derived quatities of management interest

mngVcov

An FLPar with the variance matrix of management quanties

diags

A data.frame with residuals and covariates from fit of CPUE to stock

objFn

FLPar with objective function

ll

FLPar with negative log likelihood by data component

profile

data.frame not yet implemented

hst

data.frame not yet implemented All slots in the class have accessor and replacement methods that provide validation and protection of their data.

Examples

if (FALSE) biodyn()