hcr

Harvest Control Rule, calculates F, or Total Allowable Catch (TAC) based on a hockey stock harvest control rule.

Harvest Control Rule, calculates F, or Total Allowable Catch (TAC) based on a hockey stock harvest control rule.

# S4 method for biodyn,FLPar
hcr(
  object,
  refs = hcrParam(ftar = 0.7 * mpb:::fmsy(refs), btrig = 0.8 * mpb:::bmsy(refs), fmin =
    0.01 * mpb:::fmsy(refs), blim = 0.4 * mpb:::bmsy(refs)),
  params = refs,
  stkYrs = max(as.numeric(dimnames(stock(object))$year)),
  refYrs = max(as.numeric(dimnames(catch(object))$year)),
  hcrYrs = max(as.numeric(dimnames(stock(object))$year)),
  tac = TRUE,
  bndF = NULL,
  bndTac = NULL,
  maxF = 2,
  ...
)

# S4 method for FLStock,FLBRP
hcr(
  object,
  refs,
  params = hcrParam(ftar = 0.7 * fmsy(refs), btrig = 0.8 * bmsy(refs), fmin = 0.01 *
    fmsy(refs), blim = 0.4 * bmsy(refs)),
  stkYrs = max(as.numeric(dimnames(stock(object))$year)),
  refYrs = max(as.numeric(dimnames(catch(object))$year)),
  hcrYrs = max(as.numeric(dimnames(stock(object))$year)),
  tac = TRUE,
  tacMn = TRUE,
  bndF = NULL,
  bndTac = NULL,
  maxF = 2,
  ...
)

Arguments

object
an object of class biodyn or

...
other parameters, i.e. refs FLPar or FLBRP object with reference points, can be missing if refpts are part of object params FLPar object with hockey stick HCR parameters, see hcrParam yr numeric vector with years used to values in HCR byr numeric vector with years used for bounds hyr numeric vector with years to use in projection tac logical should return value be TAC rather than F? bndF vector with bounds (i.e.min and max values) on iter-annual variability on F bndTac vector with bounds (i.e. min and max values) on iter-annual variability on TAC stkYrs numeric vector with years for calculating stock status, max(as.numeric(dimnames(stock(object))$year)), refYrs numeric vector with years for calculating reference points, max(as.numeric(dimnames(catch(object))$year)), hcrYrs numeric vector with years for setting management, max(as.numeric(dimnames(stock(object))$year)), tacMn logical should the TACs be the average of the values returned? maxF =2 numeric vector specifying maximum relative F

Value

FLPar object with value(s) for F or TAC if tac==TRUE

FLPar object with value(s) for F or TAC if tac==TRUE

Examples

if (FALSE) {
bd   =sim()

bd=window(bd,end=29)
for (i in seq(29,49,1))
bd=fwd(bd,harvest=hcr(bd,yr=i,yr=i+1)$hvt)
}
if (FALSE) {
bd   =sim()

bd=window(bd,end=29)
for (i in seq(29,49,1))
bd=fwd(bd,harvest=hcr(bd,yr=i,yr=i+1)$hvt)
}