mvn

Takes a fitted biodyn object and uses the covariance matrix (in the vcov slot) and the parameter estimates (params slot) to use Monte Carlo simulation to generate new parameters

mvn(
  object,
  n,
  nms = dimnames(object@control[object@control[, "phase", ] > 0, ])$params,
  fwd = FALSE,
  params = TRUE
)

Arguments

object
a biodyn object

n
numeric with number of iter to create

nms
the parameters for Monte Carlo simulation

fwd
logical do you want to simuate historic time series of stock biomass?, default is FALSE

params
actual parameters

Value

biodyn with simuated time series

Examples

if (FALSE) {
x=1
}