mseBiodyn

Runs a full MSE using an FLStock object as the Operating Model and biodyn as the Mangement Procedure

mseBiodyn(
  om,
  eql,
  srDev,
  uDev,
  mp,
  start = range(om)["maxyear"],
  end = start + 30,
  interval = 3,
  oem = oem,
  hcrPar = function(mp, ftar = 0.7, btrig = 0.6, fmin = 0 - 1, blim = 0.001) {
    hcrParam(ftar = ftar * fmsy(mp), btrig = btrig * bmsy(mp), fmin = fmin * fmsy(mp),
    blim = blim * bmsy(mp)) },
  bndF = NULL,
  bndTac = NULL,
  maxF = 1,
  omega = 1,
  refB = 1,
  qTrend = 0
)

Arguments

om
an FLStock object

srDev
a FLQuant with recruitment deviates

uDev
an FLQuant or FLQuants with CPUE residuals

mp
an biodyn object that holds the options for the biomass dynamic assessment model

range
a vector the starting and end years for the projections, and the interval for running the MP

ftar
a numeric with target F in HCR

fmin
a numeric with minimum F in HCR

blim
a numeric with biomass limit for HCR

btrig
a numeric with biomass trigger (i.e. when to reduce F) in HCR

what
a character that specifies what is to be used for the reference point in the HCR, recycled as required

mult
a logical that specifies whether quantity in HCR options is a multiplier or probability, recycled as required

Value

a list of data.frames with performance measures from OM and summaries from MP, if con!=NULL will also write to a MYSQL database

See also

Examples

if (FALSE) {
library(mpb)
library(FLash)
library(FLBRP)

load(om)
load(eql)

om=mpb::fwdWindow(om,eql,end=2030)
om=propagate(om,100)

srDev=FLQuant(0,dimnames=list(year=2000:2030))
srDev=rlnorm(100,srDev,0.3)

om=mpb::fwd(om,catch=catch(om)[,ac(2000:2011)],sr=eql,sr.residuals=srDev)

library(popbio)

mp=mpb::FLBRP2biodyn(  eql,"biodyn")
mp=mpb::FLStock2biodyn(om, "biodyn")
}